Beginning of koshdb
Can now save and load the password database.... Well, only the master
key for now... And the passphrase is hard coded as 'foobar'.

The database format is designed with the intent that it should be
merge-able not only by kosh, but also external diff utilities that
cannot decode the contents (such as git merge). To this end each entry
should take up a single line and be in ASCII form - conflicts will be
detected and handled by kosh the next time the database is loaded.

The first line in the file should contain a string identifying that it
is a koshdb and the version to allow detection of old databases and
migration to new formats. For the moment this is 'K05Hv0 UNSTABLE' to
indicate that it is in a state of flux and is not yet supported for
backwards compatibility.

Master key lines begin with k: and are base64 encoded. Typically there
would only be one per file, though this is not a restriction of the
format and is intentional to allow external programs to merge databases
even if the key is changed. Later this may be expanded for multiple
users, such as in a sysadmin environment, though this is not a priority.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>