Add rudimentary curses prompts and master key checksum
This adds the beginnings of the curses interface in the form of a
password prompt dialog. It's pretty rudimentary at this stage - no
colour as yet.

Two urwid widgets are added - an enhanced text box that responds to ^u
and a password input that masks it's characters.

Additionally, the SHA256 of the master key is now stored encrypted along
with the key to allow verification of the correct master passphrase.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>