Don't use weakref for timer callback and kill timers on exit
Seems that using a weak proxy object doesn't actually work in this case
as I had hoped, so scratch that idea and go back to having the database
manually expire every key when it is destructed.

Solves the program not terminating instantly on exit and the invalid
weak reference errors.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>