Add utf8 console graphics
Adds the necessary mappings from utf8 names to the sequence so that
symbols can be referred to by name in code. Allows Dividers and
SolidFills to easily be created for a BOX DRAWINGS LIGHT VERTICAL or
similar.

Not all definitions work and I don't expect many of them to - but I
don't yet have a thorough understanding of it all.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>