use entry.name instead of entry['Name']
I had intended the name of a password entry to be separate from it's
data, but in several places had started referring to it as data.

This is really a matter of policy - either way could work, but the
intent is that the name is used as a key so it makes sense to keep it
separate.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>