Add password history support
When password entries are modified, rather than simply overwriting the
old entry like every other password manager on the planet we actually
want a *new* entry to be created with the same name.

Doing so means we can retrieve older versions of passwords in the case
that we may require them again (password used in multiple places and not
all updated, old backup with old passwords restored, ...). It also gives
us better piece of mind that we won't accidentally wipe out an old
password.

It also provides an audit trail, which could be useful when password
databases are shared between multiple users (of course if the user can
edit the database directly they can do what they like, but in the future
I may well add network, daemon and/or suid support and it may well be
that some users can only access the database through kosh itself).

Entries are timestamped to sort their order out, showing only the most
recent by default (f5 will show older versions and f6 will show newer).
Currently no support exists for deleting an entry, but the idea is that
a placeholder would be used with only the name and timestamp set rather
than actually deleting an entry. I have no immediate plans to provide a
real delete/purge function.

As yet there is no detection for bogus timestamps - a good idea may be
to check for timestamps in the future when loading a koshdb and prompt
to correct them to present time then, but I want to think on that some
more first.

This patch also changes how the database is loaded and saved, partially
to accommodate the password history requirement, but also to allow
un-decodable lines to be saved and line order preserved. Practically
this allows for the inclusion of comments in the database, but also
leads us closer to multi-user support (in which not all users will be
able to decode all entries).

Finally, the structure of the password entries within the database is
changed by this commit, to include the timestamp and metadata.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>