Apply sensible ordering to fields in password entries
The order that the fields are presented in a password entry has some
importance. By default we should place Username first and Password
second. It's also foreseeable that a user may wish to place a specific
ordering on some entries.

This patch introduces a default ordering of the fields, which can be
overridden if by the FieldOrder metadata if present.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>