X Integration: WIP Initial test
This is the initial test for integration with the X clipboard via
python-Xlib. Currently this is a standalone script, but once working it
will be integrated into kosh.

The basic process according to the ICCCM is:

* Open a connection to the X server on the given display
* Create a window (not visible, minimum dimensions are 1x1)
* Generate a timestamp by synthesizing an X event (I do not know if the
  method I have used to do this is appropriate or even valid, but it
  seems to work).
* Take ownership of the desired selections with that timestmap
  - Selections include PRIMARY, SECONDARY and CLIPBOARD, each of which
    may be used in different circumstances (middle click, ^v,
    shift+insert, edit->paste, ...).  Note that the CLIPBOARD selection
    is used by clipboard managers such as klipper, we should reject
    requests for the CLIPBOARD from such programs (not yet implemented).
* Wait for a selection request, and reply appropriately (actively
  rejecting the request if timestamps are out of bounds, we don't own
  the requested selection, we don't support the requested format, etc).
  The ICCCM specifies some selections types that we must implement
  (which we don't yet), including TARGETS and TIMESTAMP.
* Update the property specified in the request on the remote client.
* Express interest in a PropertyNotify event for the requestor window.
* Send a SelectionNotify event to indicate the property has been
  updated.
* Await confirmation that the property has been changed before
  discarding the selection.

Note that it is valid to respond to a request for a selection that we do
not currently own, if the timestamp in the request is for a time that we
did own the selection, however this is generally unecessary.

If anything fails, reject the request by replying with a PropertyNotify
with an empty property.

At the moment we do not completely implement the requirements of the
ICCCM and will break in most circumstances - see the various FIXMEs and
TODOs in the patch.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>