X Clipboard WIP: Cleanup and select support
This patch cleans up the X clipboard code a little, refactoring
some functionality into standalone functions.

We also try to use select to determine when we have receieved an X
event, so that we can later respond to keyboard input. Select turned out
to be a little difficult, and I'm not convinced that we don't have a
race condition with this solution, but so far it seems to be working.

The issue was that calling select when there was already an X event
waiting would not immediately return the event. Processing every pending
X event after select returns before calling select again, but my concern
is what happens if an X event arrives after the while
display.pending_events() and before the select() call...

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>