X Clipboard: Complete clipboard integration via Xlib
This patch adds the remaining support required by the ICCCM to integrate
with the X clipboard.

This also changes the default behaviour to take control of the CLIPBOARD
selection as well as the PRIMARY and SECONDARY selections, since some
apps use the CLIPBOARD selection exclusively when pasting in certain
ways (such as ^v in Firefox).

Some programs (such as Firefox) do not expect the clipboard to change
immediately after pasting, so this waits a short time (1/10th of a
second) for any additional X events before changing the selection. This
does mean that pasting twice in very rapid succession will get the same
output.

Another approach that might work if the same timestamps are used for
additional requests of the same selection is to check the timestamp on
the SelectionRequest and return the appropriate clipboard contents for
that time. I'd need to verify Firefox' behaviour, but I think it uses a
new timestamp on the second request, in which case that approach would
not work. Regardless, the delay would still be necessary before
returning and destroying the window.

This still does not integrate with kosh, but that is now pretty trivial.

I haven't tested requesting the timestamp that the selection was
required. So far I have not come across an application that uses this,
so I can't be certain that I got it right.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>