X Clipboard: Integrate into kosh
Finally, we add support for the X clipboard into kosh itself. This adds
a clipIter to each passEntry that returns an iterator suitable for the
order of fields to be placed in the clipboard. By default it returns
[ 'Username', 'Password' ] to copy the username first and then the
password when that has been pasted.

This does do a bit of a hack to cast the blob being pasted into a string
- I was getting some issues with utf8 strings in the database not being
supported by the STRING taget. I should return them as is if a
UTF8_STRING is requested (which many apps do request first). I should
also check to see if I actually need to cast it first.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>