ssh password changer prototype: 1st commit
I'm working on a script to automate changing passwords over SSH. For the
moment I'm implementing it in a separate script, but once it works I'll
look into integrating it into the main password manager.

Actually integrating it will require some additions to the password
database format - I want to have each machine have it's own entry that
is updated independently of the password entry. The password entry will
point at all machines that should have that password, and that machine
entry will point to the password that it is currently set to:

- We update a password entry (ie, create a new entry with a new timestamp)
- This entry points to all machine entries that the old entry pointed to.
- The machine entries still point to the old password entry
- As the password is changed (and confirmed) on each machine, it is
  updated to point to the new password entry.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>