CSV Import support: Basic support
This patch adds support for importing CSV files into the password
database. Currently this is in the form of a standalone script which
opens the database and imports the entries in the file asking questions
on the command line, rather than being called from within the main
interface (I want to clean up the UI before I bother tying this in since
is it not a common operation).

Records will only be imported if a matching entry is not already found
in the database (disregarding the timestamp on the entry, which we always
expect to differ). Note that this implies that if an entry has
previously been imported and then changed and imported again, it will
still not be reimported, as there is still a matching entry in the
database.

Headers can be specified on the command line, read from the CSV file if
present, or interactively entered (which will show a example of the data
for that field, and more examples can be requested if desired). Some
TODOs in this area include testing CSV files with headers, confirming
passed in headers, not crashing if the wrong number of headers are
entered, and improving the show next example data such that duplicate
entries won't be shown many times (think categories).

A name is currently required on the imported entries and an exception
will be raised if a name does not exist on an imported entry. To solve
that issue, I need to confirm if no headers are named 'name' and prompt
the user for a name for each entry without a name... Or possibly
generate a UUID and leave the entry nameless, though that would not be
my preference.

The milestone of 'Allow myself to begin eating my own dogfood' has now
been reached.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>