SSH Password changer: Workaround pxssh bug
I have found a system with a MOTD that was too _simple_ for pxssh to
recognise. This replaces the synch_original_prompt function with
something that seems to work for all the servers I've tried so far.

After successfully implementing this I went to submit a bug report to
pxssh, only to find that someone else has made a very similar fix to
this and already submitted it, so hopefully I won't have to override
this in future.

TODO: Version check pxssh to determine if the override is necessary (for
the momemnt it is necessary in all versions).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>