curses: Introduce vi style status, command & search bar; cleanup
This introduces a new WIP widget which adds vi style functionality,
including normal, command and search modes. It is intended to be used as
a top level frame style widget which can intercept keyboard input for
processing in other modes.

It still needs improvement - for instance it has no insert mode and will
intercept keys even when a textbox is selected. I need to give that case
some thought.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>