kosh: Change default database location
Update the default database location to ~/.kosh/koshdb. This has the
advantage that we can use git (or similar) to track changes to the
database (not so much for the history since the database timestamps all
modifications) and, more importantly, to synchronise the database
between different machines.

Due the way in which the database is designed, it is perfectly valid to
use git to resolve conflicts between different databases, though I
should note that the highly recommended way to resolve such a conflict
is to append both versions and allow kosh to figure things out when it
reloads it.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>