viCommandBar: Add ability to set and display variables
This patch adds the ability to display and set the value of variables
from the viCommandBar with the 'set' command. Like vi it sets booleans
like 'set incsearch' and unsets them with 'set noincsearch'. Strings are
set like 'set somevar=foo bar' and unset with 'set somevar='. Variables
can only be set if they exist and can only be set to values of the same
type. 'set' will display all variables and 'set variablename' will
display the value of a non-boolean variable.

This still needs some cleanup, and we don't display the variables very
well just yet, but it's good enough for now.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>