X clipboard: Clean up printing
The Xclipboard module was printing debugging and status out to the
console while I developed it. Now it works we need to clean up the
messages it prints, and send them through the proper UI.

I still need to update the UI while in this code, otherwise urwid's main
loop is frozen and we will never see the messages until after the
clipboard stuff is done.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>