X clipboard: Integrate with UI main loop, fix XA_TARGETS
Fix up the clipboard integration such that it will handle urwid stuff as
well, - keeping the window drawn with up to date status. Process
keyboard events to allow blobs to be skipped and allow the process to be
cancelled via the escape key. Complete passing status to urwid widgets.

I have a feeling I may have introduced a bug when I did this, but I
can't see anything wrong. The selection ownership seemed to remain with
kosh after destroying the window. Closing the connection to the X server
did the trick, but I would still like to know why this behaviour
changed.

Also, fix an issue with the XA_TARGETS target. I had a feeling this
might turn out to be wrong, but the docs were rather incomplete with how
to pass a list through python-Xlib. I'm still not convinced that it is
correct, but this seems to have resolved an issue with iceweasel 3.5.15
pasting with ^V.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>