HTTP password changer: Rapid prototype first commit
Similar to the SSH password changer, I'm implementing a way to change
passwords on websites, by recording the sequence of actions necessary to
change the password, then playing it back.

I will *not* be adding a javascript interpreter (are *you* comfortable
with a js interpreter in your password manager, executing arbitrary code
from teh internetz? I know I'm not), so this will not work on all sites.

Like the SSH password changer I'm prototyping this in a separate script
to allow me to hack it together and debug as quickly as possible, and
will integrate it later.

The same changes to kosh I mentioned in the first SSH password changer
commit will be required before I can integrate this.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>