n900: Improve replacement PRNG
Originally I wrote it to generate a long integer of arbitrary length, however
as the python-crypto random API returns a string, I have modified it to do the
same. This reduces the amount of calls to random in favour of extra logic.
Probably didn't save anything over mapping random.uniform to a list of the
correct size...

My bigger concern is that we may not be using a high entropy source - if
python uses /dev/[u]random it should be fine, but on platforms where it falls
back to the current time my concern is that an attacker who can observe the
timestamps on the database may be able to guess the random data produced,
giving them a great deal of data they need to crack the database.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>