X Clipboard: Improve status messages
Instead of stating that we are ready to send 'credentials' by clipboard,
say the actual field that we are about to send.

Also remove the 'skipping' and 'cancelled' messages as they are both
redundant.

Finally, ensure that we print out the text name of the atoms we are
using for the clipboard, rather than their number. I don't resolve it
back into a string, so passing in Xatom.blah would still print out a
number.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>