SSH password changer: fix breakage, cleanup, robustness++
While converting this to use a cleaner interface than what pexpect
usually uses (instead of numbers for each possible match, just work on
lists of different categories and return the list itself for the match),
I had broken the script. This fixes it.

Also cleans it up a great deal and makes it more robust.

Output is now filtered such that the new password is also substituted in
the log. For the console version I make much more use of coloured output
to make it easy to see which output comes from where.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>