HTTP password changer: A whole bunch of infrastructure
We now have infrastructure to record scripts navigating HTML web pages.
At the moment we have actions to 'goto' a URL, go 'back' in history,
'undo' a mistake and 'quit'.

History is kept in the state, including actions, URLs, response info and
HTML body.

Next is to add actual HTML parsing to follow links and fill in forms,
then track cookies and add basic and digest authentication handlers.
Then actually save the script (which should be trivial) and replay it
(which will be almost as trivial so long as nothing changes ;).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>