HTTP password changer: Address unicode issue + improve code
We now catch a unicode exception when printing in ui_tty and try
encoding the string in latin-1 and re-printing. This should solve a lot
of the unicode issues, although it does make a huge assumption that the
8 bit string is latin-1.

Also improve the link handling code - some of this will be reused by the
form handling code so make the Link class do more to make the code more
generic.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>