HTTP Password Changer: Handle forms, referer, user-agent, view source
This adds the ability to fill in and submit forms. The form interface
could still use some polish, but it's enough to fill in most forms.

Replaying the form does not currently verify that the form is on the
page and is correct - it just submits the values regardless.

Still need to be able to follow links/fill in forms that have multiple
identical copies on the page.

Also should filter out forms with no submit button or action URL.
Possibly I might allow the user to specify where to submit them?

This patch also adds support for referer URLs - which default to the
previous URL (unless our current state has use_referer = False, such as
goto), but can be overridden in the script.

This also sets a kosh specific user agent and allows it be to overridden
in the script.

Finally, we add a command to view the headers and source from the reply.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>