HTTP Password changer: Finish form submitting functionality
Now when submitting the form we no longer just pass in the parameters we
got earlier when asking the user (that would not suit replays). We now
parse the document looking for the form again and take the default
values from that, only overriding whatever was passed in.

We now also support prompting for username & passwords for the
interactive case. When called from kosh we would expect them to be
passed in to us.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>