HTTP Password changer: Track cookies & allow resetting credentials
This is the first draft of cookie support. At the moment cookies won't
be undone on the undo command.

This also prompts the user if they want to enter a different value for
the username, old or new password if they fill a form in with one and
had already set it.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>