HTTP password changer: Allow credentials to be pasted in for convenience
While testing this I'm typing in the same credentials over and over and
over... This let's me just paste them in instead. They are encoded so
people can't see them from looking over my shoulder, but this is not
intended to be a secure solution, just a temporary hack until I get all
this nicely integrated into kosh.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>