HTTP Password changer:  Convert JSON loaded objects to str
Replaying the scripts kept winding up with that damn ordinal not in range
bs as it forced a link on the web page to be converted into unicode
without knowing what it was encoded in.

Recursively transform the result of the JSON load into ordinary strings
to avoid this mismatch.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>