HTTP Password changer: Allow adding additional fields to submitted forms
Bad HTML is a nasty curse, but this can help work around it. If some
fields are missing from a form (say unclosed input tags), we now have
the option to manually add the additional fields back into the form.
Manually added fields will not be checked for existence when the form is
submitted, for obvious reasons.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>