HTTP password changer: Add exact match, improve error handling, cleanup
Links can now be matched exactly by searching for "Link" (with the
double quotes) instead of link, which might match too many links. I
still don't have a way to match links with duplicated names.

Also improve the HTML parsing in terms of errors - if feed appears to
succeed, call close to flush the buffer, it turns out it didn't always
do that if it hit certain types of garbage.

Always display the script stack before asking for the next action - I
was sick of it scrolling off screen if I hit a couple of no-op keys.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>