HTTP: Add validate, handle forms w/o action, catch ReplayFailure
We can now match arbitrary text on a page to validate that we are where
we are supposed to be.

ReplayFailure is now caught, if interactive we will be prompted for the
next action again, if not interactive the script will abort.

Properly handle forms without an explicit action - they submit back to
themselves, which is handled magically by urljoin(x, None), but I had a
couple of places where I was still calling form['action'] instead of
form.getaction()

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>