X clipboard: Display status while connecting to X
On the n900 it takes a considderable amount of time to import Xlib and connect
to the display, so rather than directly importing Xlib from the xclipboard
module, delay importing until the sendviaclipboard routine is called after
displaying a status message.

Also rather than raising an XlibNotFound exception just set a status and
return.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>