UI TTY: Make ttycolours more generic
Change the style of the ttycolours to parse a given colour one word at a
time, translating them into the ANSI escape sequence. This does away
with all the 'dark' colours, and turns all the non 'dark' colours into
'bright' colours, which should be more consistent than before.

For the moment I first reset the colours to white on black as most of
the places I have used colours assume the background is black and the
colours won't look good on other backgrounds. Possibly in the future I
could abstract this away in a theme, like the urwid pallets.

Also add a _ctext_escape_len() function which will return the length of
the ANSI escape sequences used by a given colour - add this to any field
length while printing to ensure the text lines up as expected.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>