HTTP Password changer: Don't store unnamed submit buttons
If a particular submit button was selected while filling in the form, we
would save that button to be set again on replay -- which is desirable
if there are multiple submit buttons in a form.

I missed a check to make sure the submit button had a name associated
with it, so replaying would fail when it discovered there was no Null
elements in the form.

If the submit button has no name, don't have it.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>