http password changer: handle changed realms
Replaying the script would use the realm orignally recorded for
authentication, and would fail to authenticate. If the realm changes it
would typically be invisible to the user, so print out a warning and use
the realm the website presented.

If the website did not present a realm just use what was in the script
-- in case the particular pages protected by it changed and nothing
else.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>