ssh password changer: Allow for higher latency while synching prompt
We already override the synch_original_prompt function from pxssh. It
seems that when trying to log into a system on the other side of the
world, the 0.5 seconds latency it had allowed for was only sufficient if
both ends were on a fast connection.

If the first attempt to sync the prompt fails, retry allowing for a
higher latency - that way we won't slow down logins to closer and faster
systems.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>