ssh password changer: override pxssh login function
The pxssh login function does not 'your password will expire in ...'
messages, as it matches 'password' and assumes that it is at a second
password prompt and has failed to log in.

We override the login method and add an explicit check for that message,
continuing if we see it. This does not affect the behaviour if it really
is a second password prompt.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>