X clipboard: Display requestors & add blacklist
When something requests the clipboard contents, walk the window tree to
locate the client window so that it can be displayed in the interface.

This also allows us to determine if the requesting application has been
blacklisted and deny it's requests. Typically we would blacklist
clipboard managers which would subvert our control of where the
credentials are send, and perhaps even save the credentials in a
clipboard history.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>