X clipboard: Handle _NET_WM_NAME & WM_COMMAND from client leader
The Nokia N900 uses _NET_WM_NAME (of type UTF8_STRING) instead of
WM_NAME, so also look for that property.

Only one top-level window of a given client contains the WM_COMMAND
property, which may not be the top-level window on the hierachy we are
walking. Resolve the WM_CLIENT_LEADER property to find the leader window
for the client and see if it has a WM_COMMAND property.

I don't see a requirement in the ICCCM that the client leader is
necessarily the window with the WM_COMMAND property, so this may not be
sufficient for all cases. Certainly the name or commandline of the
clipboard manager on the N900 is not being resolved currently and this
is one possible explanation. So far, I haven't figured out how to
iterate over the other windows of the client to find it.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>