X Clipboard: Don't abort if lost control of PRIMARY
If we lose control of PRIMARY it was likely just been due to an implicit
selection of some text, so continue with any other selections we still
control.

If we lose control of CLIPBOARD we know it was an explicit action from
the user that caused it, so still abort in that case.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>