X Clipboard: Send subsequent credentials through same selection
Once we get a selection request and send the first credential, we almost
always expect subsequent requests to come through the same selection.

This may not be universally true... but for the moment I'm going to only
allow subsequent requests to go through the same selection anyway. If
this becomes an issue feel free to revert this patch.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>