DB: Store meta as dict (WARN: Changes db format)
Metadata was accidentally being stored as a list, but should have been a
dict. Correct new entries.

WARNING: This will change the database format for new entries!
Since we don't actually store metadata yet, this is trivial to fix - but
I'm not going to automate the update this time since the db format is
still marked as unstable and I am the only user.

Failure to fix the db won't cause issues yet, but will in the next
patch.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>