Add deletion and rename support
When entries are renamed we add the metadata RenamedFrom to point to the
old entry, so we still have a history for the renamed entry.

Deleted entries simply consist of a new entry with a name and
Deleted=True metadata.

Note there are some edge cases related to the history I haven't handled
yet - renaming one entry over another would not display one side of the
history, and if the entries are out of order in the db the delete entry
may not get applied to the correct entry (there is also a more generic
edge case of ordering in general).

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>