SSH password changer: Allow even longer timeout for prompt synch
I'm still running into timeouts when changing passwords in germany from
a mobile broadband connection over the other side of the world on a VPN.

This patch will cause the prompt synch routine to keep trying a bit
longer if it doesn't succeed at first (will now wait 3, 5 & 10 seconds
for the prompt as well as the 0.5 & 1.5 seconds I already had it waiting
for).

Also, set login_timeout to 30 seconds, which makes it less susceptible
to being tricked if it takes a while for the password prompt to show.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>