SSH Password changer: Make password verification more reliable
It seems that the login stage of pxssh can sometimes believe that it has
successfully logged in if the password prompt takes too long to display
(perhaps assuming it is passwordless authentication?).

Regardless, this patch makes us actually try to do something on the
remote system and check the result to verify that we have indeed logged
in. We don't simply echo something as that may be echoed back to us and
we would get a false positive, instead we check the result of some small
arithmetic which can't be fooled by an echo.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>