Revert "X Clipboard: Send subsequent credentials through same selection"
Ok, so I've found one reasonably common case that does not fit this
idea:

In firefox if a username text input box is selected that immediately
opens a dropdown of remembered usernames, pressing shift+insert will use
the CLIPBOARD selection, instead of the PRIMARY selection that it
usually uses (likely somehow related to the fact that firefox grabs the
input when a dropdown is displayed, and may therefore handle
shift+insert differently to the standard X method).

Then, when the password field is selected (which won't have a dropdown
since firefox isn't supposed to remember password fields in the same
way) shift+insert will trigger the standard X method which uses the
PRIMARY selection.

This gives us a reasonably common case where the username and password
may be requested via different selections, so revert this patch.

This reverts commit 02f230f884bb217a210d1085edddd6ba19e2988d

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>