Hack to support generating passwords
Pressing ctrl+g on a password field will call out to pwgen to gnerate a
12 digit password.

The code isn't in the most appropriate place at the moment, and makes a
lot of assumptions.

This also removes the python2.6 hack (and assumes it is run on python
2.7...), so will break on prior python versions - though that falls unde
the n900 effort to ensure compatibility with python 2.5 anyway and all
breakages represent bugs.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>