Fix input handling breakage
Removing the python2.6 version a few patches back meant that we now use
python2.7 on most modern systems.

Unfortunately, that broke the (hacky) input handling in the X clipboard
code. It seems that it is to do with the file object sys.stdin being in
the list of FDs that select was waiting on in addition to 0, the fileno
of sys.stdin.

Since stdin is going to be one of the fds that urwid selects on, and we
already add those to the select loop in the X clipboard code, don't add
sys.stdin and test for it's fileno.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>