n9: Fix strxor package vs function on N9
It seems that strxor does exist in python-crypto on the N9, but it is now a
function in a strxor package in the Util package, rather than a function
directly in the Util package (I'm getting a little alarmed at the apparent
instability of the crypto package utilities API...).

This patch adds a test to see if strxor contains another strxor and uses
that if it exists.