Make clipboard exceptions / bugs non-fatal
Exceptions in the clipboard integration are annoying, but are fairly
self-contained and shouldn't result in kosh dying and having to be
restarted.

Expected exceptions (like, couldn't connect to X display) are handled
gracefully by displaying a message in the statusbar.

We show a traceback if an unexpected exception occurs, but we then
resume kosh rather than exiting. I want to add similar handling in other
areas as well (particularly calls to the HTTP/SSH scripting).

One issue I've noticed is that after displaying an exception all the
non-ASCII characters are wrong. This also occurs after calling out to
the HTTP script. I suspect that this is because I'm instantiating a new
urwid main loop rather than using the existing one (seemed easier at the
time to make it modal)... I need to figure out how to do this properly
at some point.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>