N9: HACK X Clipboard code until it works
I'm... disappointed.

The N9 employs a clipboard manager. So far nothing wrong...

The clipboard manager on the N9 aggressively controls the CLIPBOARD
selection, even if the CLIPBOARD selection owner refuses the
XSelectionRequests of the clipboard manager (as any good password
manager should do). This behaviour is both rude and inappropriate.

It appears that the clipboard manager is *mandatory* (if enabled?) - N9
apps do not appear to use the selection method themselves, so we cannot
pass the credentials directly to them. This is bad.

Well, that's not entirely true - it appears that the foreground app
*does* request the CLIPBOARD as TEXT at the same time that the clipboard
manager. This is ineffective if the clipboard manager is active (though
that is not a reason for it not to). This is the wrong behaviour even if
the clipboard manager wasn't active - it should request it as TARGETS
only until it is ready to actually paste the CLIPBOARD contents.

After connecting to the display, creating a window and taking ownership
of the CLIPBOARD selection, the clipboard manager (and foreground app,
if available) immediately requests the selection. This is normal for a
clipboard manager. It is legal for the foreground app, but would usually
only be appropriate if a UI element is visible that needs to provide
feedback as to whether pasting is possible (ie, greying out/activating a
toolbar button), which is not the case here. This may be mandatory for
the QT framework.

After the clipboard manager has forcefully taken control of the
CLIPBOARD selection, it ignores any future times that the window takes
ownership of the selection. This makes no sense at all.

I might need to see if there is any way to do the right thing with the
QT clipboard interface. Or if I can SIGSTOP the clipboard manager.

This patch:

- Adds a bunch of optional debugging info
- Fixes the TIMESTAMP target so python-Xlib doesn't raise an exception.
- Checks if the display has any events before using select. This is to
  minimise the chance of hitting a race condition when using select on
  python-Xlib. Unfortunately this does not eliminate this race (this is
  not specific to the N9, it just happens to always hit on the N9
  without this due to the clipboard manager).
- Does not give up when the CLIPBOARD selection is lost (since we expect
  it to be when the clipboard manager takes control of it), rather waits
  for the user to press enter/escape.
- Destroys and recreates the connection to the display on each
  individual blob to workaround the clipboard manager only taking note
  the first time we take ownership of the CLIPBOARD selection.
- Unregisters interest in property updates once a selection transfer is
  complete.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>