N9 X Clipboard: Hack until it does the right thing
Ok, this is nasty and needs more testing.

On the N9, if we see the clipboard manager request a selection, we not
only deny it, but send it a SIGSTOP to let it know just what we think of
it (we do sent it a SIGCONT before we leave - we aren't cold hearted,
just hackers).

With the clipboard manager out of the way we can now deal directly with
the applications. yay :)

Unfortunately they are just as broken as the clipboard manager and
request the contents of the CLIPBOARD as TEXT straight away (four
times even!!!!), so... we can't use that to determine when to switch to
the password like we would in a sane world :(

But, all is not lost! See, most apps seem to clear the clipboard after
using it on the N9 (why?) or provide an easy mechanism to do so, and we
get notified when that happens! So, we can use that to then place the
password in the CLIPBOARD :)

The apps are still brain-dead - like the clipboard manager they ignore
my window taking the CLIPBOARD selection a second time, so I still have
the re-connect to X hack in place (yuck).

Seriously, what the hell is going on in the QT clipboard framework?
Whatever it is, it ain't natural, I only hope there is a good reason (I
can think of a couple of reasons, all of which could have been solved in
better ways like the rest of the world does, I just can't think of a
good reason for doing it this way... My guess - a lack of understanding
of the intricacies of the ICCCM).

This needs wider testing on the n9 and elsewhere, and I still need to
implement the n9() stub to control which behaviour we use.

Signed-off-by: Ian Munsie <darkstarsword@gmail.com>